SPMD Control {pbdMPI}	R Documentation

Sets of controls in pbdMPI.

Description

These sets of controls are used to provide default values in this package.

Format

Objects contain several parameters for communicators and methods.

Details

The elements of .pbd_env$SPMD.CT are default values for controls including

Elements 	 Default 	 Usage

comm 	 0L 	 a communicator index

intercomm 	 2L 	 an inter communicator index

info 	 0L 	 an info index

newcomm 	 1L 	 a new communicator index

op 	 "sum" 	 an operation

port.name 	 "spmdport" 	 an operation

print.all.rank 	 FALSE 	 if all ranks print message

print.quiet 	 FALSE 	 if print/cat rank information

rank.root 	 0L 	 a rank of root

rank.source 	 0L 	 a rank of source

rank.dest 	 1L 	 a rank of destination

request 	 0L 	 a request index

serv.name 	 "spmdserv" 	 a service name

status 	 0L 	 a status index

tag 	 0L 	 a tag number

unlist 	 FALSE 	 if unlist returning

divide.method 	 "block" 	 a way to divide jobs or data

mpi.finalize 	 TRUE 	 if shutdown MPI

quit 	 TRUE 	 if quit when errors occur

msg.flush 	 TRUE 	 if flush message for comm.cat/comm.print

msg.barrier 	 TRUE 	 if barrier message for comm.cat/comm.print

Rprof.all.rank 	 FALSE 	 if call Rprof on all ranks

lazy.check 	 TRUEE 	 if use lazy check on all ranks

The elements of .pbd_env$SPMD.OP list the implemented operations for reduce() and allreduce(). Currently, four operations are implemented "sum", "prod", "max", and "min".

The elements of .SPMD.IO are default values for input and output including

Elements 	 Default 	 Usage

max.read.size 	 5.2e6 	 max of reading size (5 MB)

max.test.lines 	 500 	 max of testing lines

read.method 	 "gbd" 	 default reading method

balance.method 	 "block" 	 default load balance method

where balance.method is only used for "gbd" reading method when nrows = -1 and skip = 0 are set.

The elements of .pbd_env$SPMD.TP are default values mainly for task pull including

Elements 	 Default 	 Usage

bcast 	 FALSE 	 if bcase() objects to all ranks

barrier 	 TRUE 	 if call barrier() for all ranks

try 	 TRUE 	 if use try() in works

try.silent 	 FALSE 	 if silent the try() message

See task.pull() for details.

Author(s)

Wei-Chen Chen wccsnow@gmail.com, George Ostrouchov, Drew Schmidt, Pragneshkumar Patel, and Hao Yu.

References

Programming with Big Data in R Website: http://r-pbd.org/

[Package pbdMPI version 0.3-1 Index]