global Rprof {pbdMPI}	R Documentation

A Rprof Function for SPMD Routines

Description

A Rprof function for use with parallel codes executed in the batch SPMD style.

Usage

comm.Rprof(filename = "Rprof.out", append = FALSE, interval = 0.02,
           memory.profiling = FALSE, gc.profiling = FALSE,
           line.profiling = FALSE, numfiles = 100L, bufsize = 10000L,
           all.rank = .pbd_env$SPMD.CT$Rprof.all.rank,
           rank.Rprof = .pbd_env$SPMD.CT$rank.source,
           comm = .pbd_env$SPMD.CT$comm)

Arguments

filename
as in Rprof().

append
as in Rprof().

interval
as in Rprof().

memory.profiling
as in Rprof().

gc.profiling
as in Rprof().

line.profiling
as in Rprof().

numfiles
as in Rprof().

bufsize
as in Rprof().

all.rank
if calling Rprof on all ranks (default = FALSE).

rank.Rprof
rank for calling Rprof if all.rank = FALSE (default = 0).

comm
a communicator number.

Details

as in Rprof().

Author(s)

Wei-Chen Chen wccsnow@gmail.com, George Ostrouchov, Drew Schmidt, Pragneshkumar Patel, and Hao Yu.

References

Programming with Big Data in R Website: http://r-pbd.org/

[Package pbdMPI version 0.3-1 Index]