SPMD Control Functions {pbdMPI}	R Documentation

Sets of controls in pbdMPI.

Description

These sets of controls are used to provide default values in this package.

Author(s)

Wei-Chen Chen wccsnow@gmail.com, George Ostrouchov, Drew Schmidt, Pragneshkumar Patel, and Hao Yu.

References

Programming with Big Data in R Website: http://r-pbd.org/

See Also

.pbd_env.

[Package pbdMPI version 0.3-1 Index]